Initial commit: symbolic arithmetic
Includes the initial functionality + tests for basic symbolic
arithmetic: +, -, *, /.

All of this is included in a single commit because I copied this logic
over from its original .exs file. This problem is a bit too large/messy
to be completed in a single file, so a mix project makes more sense.

I'm not sure about my use of function overloading here. Since the
functions and conditions are very short, maybe `case` would be better?

Next up: Derivatives!